requirements.txt
fastapi
fastapi-azure-auth==4.1.3
uvicorn
sqlalchemy
psycopg2
pymysql
python-dotenv
httpx
faker
coverage
pytest
pytest-asyncio
pytest-cov
black
isort
mypy
ruff
pydantic==1.10.2
python-jose[cryptography]==3.3.0